Add "isClient" check to frontend observability
Nextjs will try to render client components on the server. The TracingInstrumentation uses zonejs and can cause issues with how next handles streaming.

In the app router specifically, if a server action was used to set or modify cookies, the resulting RSC payload would fail with this setup. After many, many hours of trying to track down why nextjs was throwing a "cookies() was not part of the request context error", it turned out it was the TracingInstrumentation() messing with the RSC payload and causing issues with AsyncLocalStorage. Adding the "window !== undefined" fixed the issues.